Made some reasonably big changes to the physics.
Changed around the log types
Implemented a PhysicalPlayer Class
Made some more generic rigidBody additions.
The storing of rigidBodies and collisionBodies still needs
to be written to not leak pointers all over the place. And we
need to get a good way of using a hash map of some sort to
store physics objects by some relevant ID. There is a getOrigin
function implemented for the player class which can be used to at
least draw the falling box without the rotations and such.